requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
langchain-openai
langchain-groq
langchain-text-splitters
faiss-cpu
python-dotenv
pypdf
sentence-transformers
